Consolidate unit tests
Remove duplicate integration tests.
